test(canon): move the virtual-root symlink tests into their own module
